+++ src/core/steps/metadata.step.js
@@ -18,12 +18,12 @@ function _camelToSnake(string) {
 }
 
 function metadataStep(pa, model, nextSteps) {
-    model.setProperty('event_collection_platform', BUILD_BROWSER ? 'js' : 'js-browserless');
-    model.setProperty('event_collection_version', model.getConfiguration('version'));
+    model.addEventsProperty('event_collection_platform', BUILD_BROWSER ? 'js' : 'js-browserless');
+    model.addEventsProperty('event_collection_version', model.getConfiguration('version'));
     const date = new Date();
-    model.setProperty('device_timestamp_utc', date.getTime());
-    model.setProperty('device_local_hour', date.getTime());
-    model.setProperty('device_hour', date.getHours());
+    model.addEventsProperty('device_timestamp_utc', date.getTime());
+    model.addEventsProperty('device_local_hour', date.getTime());
+    model.addEventsProperty('device_hour', date.getHours());
 
     if (BUILD_BROWSER) {
         const manualConfig = 'isManualPageRefresh',
@@ -43,20 +43,14 @@ function metadataStep(pa, model, nextSteps) {
                 }
 
             }
-            if (_isPropertiesAbsentForEvent(pageviewidProp, model, event)) {
-                if (pa._privacy.call('isPropAllowed', pageviewidProp)) {
-                    event.data[pageviewidProp] = dataLayer.get('pageViewId');
-                }
+            if (pa._privacy.call('isPropAllowed', pageviewidProp) && model.isPropertyAbsentForEvent(pageviewidProp, event)) {
+                event.data[pageviewidProp] = dataLayer.get('pageViewId');
             }
         }
 
         try {
             const cookieCreationDate = new Date((new Date(dataLayer.cookies._pcid.fixedAt[0])).setUTCHours(0, 0, 0, 0)).toISOString();
-            for (const event of model.events) {
-                if (_isPropertiesAbsentForEvent('cookie_creation_date', model, event)) {
-                    model.setProperty('cookie_creation_date', cookieCreationDate);
-                }
-            }
+            model.addEventsProperty('cookie_creation_date', cookieCreationDate);
         } catch (e) { /* empty */
         }
         const content = dataLayer.get('content');
@@ -67,40 +61,34 @@ function metadataStep(pa, model, nextSteps) {
                     'tags': 'tags_array'
                 };
                 const propFinalName = (propContent === 'createdAt' || propContent === 'tags') ? MAP_PA_DL[propContent] : _camelToSnake(`content_${propContent}`);
-                for (const event of model.events) {
-                    if (_isPropertiesAbsentForEvent(propFinalName, model, event)) {
-                        if (pa._privacy.call('isPropAllowed', propFinalName)) {
-                            event.data[propFinalName] = content[propContent];
-                        }
-                    }
-                }
+                model.addEventsProperty(propFinalName, content[propContent]);
             }
         }
 
-        model.setProperty('has_access', dataLayer.get('userStatus'));
-        model.setProperty('device_screen_width', window.screen.width);
-        model.setProperty('device_screen_height', window.screen.height);
-        model.setProperty('device_display_width',
+        model.addEventsProperty('has_access', dataLayer.get('userStatus'));
+        model.addEventsProperty('device_screen_width', window.screen.width);
+        model.addEventsProperty('device_screen_height', window.screen.height);
+        model.addEventsProperty('device_display_width',
             window.innerWidth ||
             document.documentElement && document.documentElement.clientWidth ?
                 document.documentElement.clientWidth : ''
         );
-        model.setProperty('device_display_height',
+        model.addEventsProperty('device_display_height',
             window.innerHeight ||
             document.documentElement && document.documentElement.clientHeight ?
                 document.documentElement.clientHeight : ''
         );
         const language = window.navigator ? (window.navigator.language || window.navigator.userLanguage) : '';
         const languageSplitted = _parseLanguage(language, ['-', '_']);
-        model.setProperty('browser_language', languageSplitted[0]);
-        model.setProperty('browser_language_local', languageSplitted[1]);
-        model.setProperty('previous_url', document.referrer || '');
+        model.addEventsProperty('browser_language', languageSplitted[0]);
+        model.addEventsProperty('browser_language_local', languageSplitted[1]);
+        model.addEventsProperty('previous_url', document.referrer || '');
         if (document.title) {
-            model.setProperty('page_title_html', document.title);
+            model.addEventsProperty('page_title_html', document.title);
         }
         const eventUrlWithQueryString = model.getConfiguration('addEventURL').toString() === 'true';
         if (eventUrlWithQueryString || (model.getConfiguration('addEventURL') === 'withoutQS')) {
-            model.setProperty('event_url_full', eventUrlWithQueryString ? window.location.href.split('#')[0] : `${window.location.protocol}//${window.location.host}${window.location.pathname}`);
+            model.addEventsProperty('event_url_full', eventUrlWithQueryString ? window.location.href.split('#')[0] : `${window.location.protocol}//${window.location.host}${window.location.pathname}`);
         }
 
         try {
@@ -137,21 +125,6 @@ function metadataStep(pa, model, nextSteps) {
     } else {
         nextStep(pa, model, nextSteps);
     }
-}
-
-function _isPropertiesAbsentForEvent(name, model, event) {
-    if (model.hasProperty(name) && model.properties[name].options.events) {
-        const propertyEventsOption = model.properties[name].options.events;
-        if (propertyEventsOption.indexOf(event.name) > -1) {
-            return false;
-        }
-        for (const eventAllowed of propertyEventsOption) {
-            if (eventAllowed.charAt(eventAllowed.length - 1) === '*' && event.name.indexOf(eventAllowed.substring(0, eventAllowed.length - 1)) === 0) {
-                return false;
-            }
-        }
-    }
-    return typeof event.data[name] === 'undefined';
 }
 
 function _isDefined(variable) {
@@ -200,7 +173,7 @@ function _addUserAgentMetadata(model, ua) {
     if (_isDefined(ua)) {
         for (let i = 0; i < properties.length; i++) {
             if (_isDefined(ua[properties[i].metric])) {
-                model.setProperty(properties[i].property, ua[properties[i].metric]);
+                model.addEventsProperty(properties[i].property, ua[properties[i].metric]);
             }
         }
     }