+++ test/shared/properties.js
@@ -188,28 +188,32 @@ describe('Properties :', function () {
             });
         });
         it('Should not override values of privacy specific properties', function (done) {
-            globalPA.setProperty('visitor_privacy_consent', false);
-            globalPA.sendEvent('toto', {}, {
+            globalPA.setProperty('visitor_privacy_consent', 'myconsent');
+            globalPA.setProperty('visitor_privacy_mode', 'test');
+            globalPA.sendEvent('toto', {'visitor_privacy_mode': 'mymode'}, {
                 onBeforeSend: function (pianoanalytics, model) {
-                    expect(model.build.data.events[0].data['visitor_privacy_consent']).to.equal(true);
-                    expect(model.build.data.events[0].data['visitor_privacy_mode']).to.equal('optin');
-                    done();
+                    Utility.promiseThrowCatcher(done, function () {
+                        expect(model.build.data.events[0].data['visitor_privacy_consent']).to.equal('myconsent');
+                        expect(model.build.data.events[0].data['visitor_privacy_mode']).to.equal('mymode');
+                        done();
+                    });
                 }
             });
         });
         it('Should not override values of metadata specific properties', function (done) {
             globalPA.setProperty('event_collection_platform', '1');
-            globalPA.setProperty('event_collection_version', '2');
-            globalPA.setProperty('device_timestamp_utc', '3');
-            globalPA.sendEvent('toto', {}, {
+            globalPA.setProperty('event_collection_version', '22');
+            globalPA.sendEvent('toto', {
+                'event_collection_version': '2',
+                'device_timestamp_utc': '3'
+            }, {
                 onBeforeSend: function (pianoanalytics, model) {
-                    expect(model.build.data.events[0].data['event_collection_platform']).to.not.equal('1');
-                    expect(model.build.data.events[0].data['event_collection_platform']).to.not.equal(undefined);
-                    expect(model.build.data.events[0].data['event_collection_version']).to.not.equal('2');
-                    expect(model.build.data.events[0].data['event_collection_version']).to.not.equal(undefined);
-                    expect(model.build.data.events[0].data['device_timestamp_utc']).to.not.equal('3');
-                    expect(model.build.data.events[0].data['device_timestamp_utc']).to.not.equal(undefined);
-                    done();
+                    Utility.promiseThrowCatcher(done, function () {
+                        expect(model.build.data.events[0].data['event_collection_platform']).to.equal('1');
+                        expect(model.build.data.events[0].data['event_collection_version']).to.equal('2');
+                        expect(model.build.data.events[0].data['device_timestamp_utc']).to.equal('3');
+                        done();
+                    });
                 }
             });
         });
@@ -437,32 +441,40 @@ describe('Properties :', function () {
         });
         it('Should not override values of privacy specific properties', function (done) {
             globalPA.setProperties({
-                'visitor_privacy_consent': false,
+                'visitor_privacy_consent': 'myconsent',
                 'visitor_privacy_mode': 'test'
             });
-            globalPA.sendEvent('toto', {}, {
+            globalPA.sendEvent('toto', {'visitor_privacy_mode': 'mymode'}, {
                 onBeforeSend: function (pianoanalytics, model) {
-                    expect(model.build.data.events[0].data['visitor_privacy_consent']).to.equal(true);
-                    expect(model.build.data.events[0].data['visitor_privacy_mode']).to.equal('optin');
-                    done();
+                    Utility.promiseThrowCatcher(done, function () {
+                        expect(model.build.data.events[0].data['visitor_privacy_consent']).to.equal('myconsent');
+                        expect(model.build.data.events[0].data['visitor_privacy_mode']).to.equal('mymode');
+                        done();
+                    });
                 }
             });
         });
         it('Should not override values of metadata specific properties', function (done) {
             globalPA.setProperties({
                 'event_collection_platform': '1',
+            }, {events: ['tata']});
+            globalPA.setProperties({
                 'event_collection_version': '2',
-                'device_timestamp_utc': '3'
-            });
-            globalPA.sendEvent('toto', {}, {
+                'device_timestamp_utc': 'test'
+            }, {events: ['toto']});
+            globalPA.setProperties({
+                'device_local_hour': '4'
+            }, {events: ['to*']});
+            globalPA.sendEvent('toto', {'device_timestamp_utc': '3'}, {
                 onBeforeSend: function (pianoanalytics, model) {
-                    expect(model.build.data.events[0].data['event_collection_platform']).to.not.equal('1');
-                    expect(model.build.data.events[0].data['event_collection_platform']).to.not.equal(undefined);
-                    expect(model.build.data.events[0].data['event_collection_version']).to.not.equal('2');
-                    expect(model.build.data.events[0].data['event_collection_version']).to.not.equal(undefined);
-                    expect(model.build.data.events[0].data['device_timestamp_utc']).to.not.equal('3');
-                    expect(model.build.data.events[0].data['device_timestamp_utc']).to.not.equal(undefined);
-                    done();
+                    Utility.promiseThrowCatcher(done, function () {
+                        expect(model.build.data.events[0].data['event_collection_platform']).to.not.equal('1');
+                        expect(model.build.data.events[0].data['event_collection_platform']).to.not.equal(undefined);
+                        expect(model.build.data.events[0].data['event_collection_version']).to.equal('2');
+                        expect(model.build.data.events[0].data['device_timestamp_utc']).to.equal('3');
+                        expect(model.build.data.events[0].data['device_local_hour']).to.equal('4');
+                        done();
+                    });
                 }
             });
         });